+++ contracts/ERC1820Registry.sol
@@ -27,7 +27,7 @@
  */
 pragma solidity 0.5.3;
 // IV is value needed to have a vanity address starting with `0x1820`.
-// IV: 42393
+// IV: 87018
 
 /// @dev The interface a contract MUST implement if it is the implementer of
 /// some (other) interface for any address other than itself.